Update to run on servers

+++ src/main/java/Benchmark/SortBenchmark.java
@@ -54,16 +54,16 @@ public class SortBenchmark {
     }
 
     public void writeUnsortedKeysToSorter(Sorter sorter, int count) throws IOException {
-        long[] key = new long[4];
+        Long[] key = new Long[4];
         for(int i = 2; i < count; i++){
             if(i % 2 == 0){
                 for(int j = 0; j < key.length; j++){
-                    key[j] = (i + 1);
+                    key[j] = (long) (i + 1);
                 }
             }
             else{
                 for(int j = 0; j < key.length; j++){
-                    key[j] = (i - 1);
+                    key[j] = (long)(i - 1);
                 }
             }
             sorter.addUnsortedKey(key);
@@ -71,7 +71,7 @@ public class SortBenchmark {
     }
 
     public void writeRandomKeysToSorter(Sorter sorter, int count) throws IOException {
-        long[] key = new long[4];
+        Long[] key = new Long[4];
         Random random = new Random();
         for(int i = 2; i < count; i++){
             long rnd = Math.abs(random.nextLong());
